+++ index.ts
@@ -1,4 +1,5 @@
 export * from "./eosiolib/asset";
+export * from "./eosiolib/extended_asset";
 export * from "./eosiolib/check";
 export * from "./eosiolib/symbol";
 export * from "./eosiolib/symbol_code";